+++ NEWS.md
@@ -1,3 +1,5 @@
+* New parameter `override.config` to `load.project()` and `reload.project()`
+  allows substitution of individual configuration items.
 * Missing but required directories are now created (with a warning).
 * All logical configuration options are now stored as Boolean values instead of
   `'on'`/`'off'`.  Input values other than `'on'`/`'off'` are converted using